+++ pro/ai/mcp/client/src/test/kotlin/org/http4k/ai/mcp/client/jsonrpc/JsonRpcMcpClientTest.kt
@@ -22,8 +22,8 @@ class JsonRpcMcpClientTest : McpClientContract<Unit>() {
     override fun clientSessions() = JsonRpcSessions()
 
     override fun clientFor(port: Int) = JsonRpcMcpClient(
-        McpEntity.of("http4k MCP client"), Version.of("0.0.0"),
         Uri.of("http://localhost:${port}/jsonrpc"),
+        McpEntity.of("http4k MCP client"), Version.of("0.0.0"),
         ClientFilters.BearerAuth("123").then(JavaHttpClient()),
     )