+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 This list is not intended to be all-encompassing - it will document major and breaking API
 changes with their rationale when appropriate:
 
+### v6.33.0.0 (uncut)
+- **http4k-***: Upgrade versions
+
 ### v6.32.0.0
 - **http4k-***: Upgrade versions
 - **http4k-ai-mcp-sdk**: [Breaking] Defaulting of MCP Entity and Version names in MCP Clients. We have reordered the parameters to the constructor to